requirements.txt
requests
garminconnect
fitparse
python-dotenv